backend/pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "python-multipart>=0.0.6",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "pgvector>=0.2",
    "minio>=7.2",
    "mistralai>=1.0",
    "celery[redis]>=5.3",
    "redis>=5.0",
    "pydantic-settings>=2.1",
    "striprtf>=0.0.26",
    "python-magic>=0.4",
    "aiofiles>=23.2",
    "httpx>=0.25",
    "beautifulsoup4>=4.12",
    "langchain>=0.3",
    "langchain-mistralai>=0.2",
    "langchain-community>=0.3",
    "sse-starlette>=2.0",
    "bcrypt>=4.0",
    "python-jose[cryptography]>=3.3",
    "openpyxl>=3.1",
    "alembic>=1.13",
]
